+++ sdk/python/waddle_sdk/simulators/scene.py
@@ -70,6 +70,7 @@ _ENVIRONMENT_REVISIONS = {
     "pick_lift": ("1.1.0", "1.0.0"),
     "candy-bin-transfer": ("1.2.0", "1.2.0"),
     "chocolate-packing": ("1.0.3", "1.0.0"),
+    "shampoo-packing": ("1.0.1", "1.0.0"),
 }
 _REVISION = re.compile(r"^[0-9]+\.[0-9]+\.[0-9]+(?:[-+][0-9A-Za-z.-]+)?$")